+++ src/app/mypage/[blogId]/components/Background/index.tsx
@@ -2,7 +2,7 @@
 
 import { FileEdit } from '@/components/Common';
 import { FileThumbnails } from '../Common';
-import useEdit from '../hooks/useEdit';
+import { useEdit } from '../hooks';
 
 export default function Background() {
   const { handleBackgroundFileChange, background, backgroundFile } = useEdit();
